refactor: 같은 bounded context 도메인 패키지 합침
