+++ src/main/java/kitchenpos/menu/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menugroup.domain;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
