+++ src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.menugroup.ui;
+package kitchenpos.menu.ui;
 
-import kitchenpos.menugroup.application.MenuGroupService;
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menu.application.MenuGroupService;
+import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
