+++ src/main/java/kitchenpos/order/domain/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.order.domain;
 
 import java.util.List;
 import java.util.Optional;
